+++ source/app/service-providers/documents/index.ts
@@ -592,9 +592,12 @@ export default class DocumentManager extends ProviderContract {
       return false
     }
 
-    if (clientVersion < doc.minimumVersion) {
-      // TODO: This means that the client is completely out of sync and needs to
-      // re-fetch the whole document.
+    if (clientVersion < doc.minimumVersion || clientVersion > doc.currentVersion) {
+      // The client is completely out of sync and has to reload the document.
+      // If this happens because clientVersion < doc.minimumVersion, this means
+      // that the lost connection somehow. If it happens because clientVersion
+      // > doc.currentVersion, it means that we had to roll over the version in
+      // pushUpdates below.
       return false
     } else if (clientVersion < doc.currentVersion) {
       return doc.updates.slice(clientVersion - doc.minimumVersion)
@@ -612,6 +615,12 @@ export default class DocumentManager extends ProviderContract {
     if (clientVersion !== doc.currentVersion) {
       return false
     }
+
+    // Before applying any updates, we have to clear any potential timeout so
+    // that it does not interfere with us updating the document. Otherwise, this
+    // can lead to a faulty state where the provider cannot save the file
+    // anymore.
+    clearTimeout(doc.saveTimeout)
 
     for (const update of clientUpdates) {
       const changes = ChangeSet.fromJSON(update.changes)
@@ -639,10 +648,9 @@ export default class DocumentManager extends ProviderContract {
     // Drop all updates that exceed the amount of updates we allow.
     while (doc.updates.length > MAX_VERSION_HISTORY) {
       doc.updates.shift()
-      doc.minimumVersion++
+      doc.minimumVersion += 1
     }
 
-    clearTimeout(doc.saveTimeout)
     const autoSave = this._app.config.get().editor.autoSave
 
     // No autosave
@@ -650,12 +658,10 @@ export default class DocumentManager extends ProviderContract {
       return true
     }
 
-    const timeout = autoSave === 'delayed' ? DELAYED_SAVE_TIMEOUT : IMMEDIATE_SAVE_TIMEOUT
-
     doc.saveTimeout = setTimeout(() => {
       this.saveFile(doc.filePath)
         .catch(err => this._app.log.error(`[Document Provider] Could not save file ${doc.filePath}: ${err.message as string}`, err))
-    }, timeout)
+    }, autoSave === 'delayed' ? DELAYED_SAVE_TIMEOUT : IMMEDIATE_SAVE_TIMEOUT)
 
     return true
   }
@@ -1300,19 +1306,27 @@ export default class DocumentManager extends ProviderContract {
       return false
     }
 
-    this._ignoreChanges.push(filePath)
+    // If saveFile was called from a timeout, clearTimeout does nothing but the
+    // timeout is reset to undefined. However, implementing this check here
+    // ensures that we can programmatically call saveFile anywhere else and
+    // still have everything work as intended.
+    if (doc.saveTimeout !== undefined) {
+      clearTimeout(doc.saveTimeout)
+      doc.saveTimeout = undefined
+    }
+
+    // NOTE: Remember that we MUST under any circumstances adapt the document
+    // descriptor BEFORE attempting to save. The reason is that if we don't do
+    // that, we can run into the following race condition:
+    // 1. User changes the document
+    // 2. The save commences
+    // 3. The user adds more changes
+    // 4. The save finishes and undos the modifications
+    const content = doc.document.toString()
+    doc.lastSavedVersion = doc.currentVersion
 
     if (doc.descriptor.type === 'file') {
-      await FSALFile.save(
-        doc.descriptor,
-        doc.document.toString(),
-        this._app.fsal.getMarkdownFileParser(),
-        null
-      )
-      await this.synchronizeDatabases() // The file may have gotten a library
-
       // In case of an MD File increase the word or char count
-      const content = doc.document.toString()
       const newWordCount = countWords(content, false)
       const newCharCount = countWords(content, true)
 
@@ -1325,20 +1339,22 @@ export default class DocumentManager extends ProviderContract {
 
       doc.lastSavedWordCount = newWordCount
       doc.lastSavedCharCount = newCharCount
+    }
+
+    this._ignoreChanges.push(filePath)
+
+    if (doc.descriptor.type === 'file') {
+      await FSALFile.save(
+        doc.descriptor,
+        content,
+        this._app.fsal.getMarkdownFileParser(),
+        null
+      )
+      await this.synchronizeDatabases() // The file may have gotten a library
     } else {
-      await FSALCodeFile.save(doc.descriptor, doc.document.toString(), null)
+      await FSALCodeFile.save(doc.descriptor, content, null)
     }
 
-    // If saveFile was called from a timeout, clearTimeout does nothing but the
-    // timeout is reset to undefined. However, implementing this check here
-    // ensures that we can programmatically call saveFile anywhere else and
-    // still have everything work as intended.
-    if (doc.saveTimeout !== undefined) {
-      clearTimeout(doc.saveTimeout)
-      doc.saveTimeout = undefined
-    }
-
-    doc.lastSavedVersion = doc.currentVersion
     this._app.log.info(`[DocumentManager] File ${filePath} saved.`)
     this.broadcastEvent(DP_EVENTS.CHANGE_FILE_STATUS, { filePath, status: 'modification' })
     this.broadcastEvent(DP_EVENTS.FILE_SAVED, { filePath })